Worked: Adding transaction (for SQLite). Implemented: date & epochTime. Implementing: swipe to delete

+++ app/src/main/java/com/example/pxh612_loginapi_v2/viewmodel/AddTransactionViewModel.java
@@ -2,7 +2,6 @@ package com.example.pxh612_loginapi_v2.viewmodel;
 
 import android.content.Context;
 
-import com.example.pxh612_loginapi_v2.activity.AddTransactionActivty;
 import com.example.pxh612_loginapi_v2.repository.TransactionRepository;
 
 public class AddTransactionViewModel {
@@ -13,7 +12,7 @@ public class AddTransactionViewModel {
         transactionRepository = new TransactionRepository(context);
     }
 
-    public void addTransaction(String category, int amount, int date) {
+    public void addTransaction(String category, int amount, long date) {
         transactionRepository.addTransaction(category, amount, date);
     }
 }
